% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data_helpers.R
\name{clean_address}
\alias{clean_address}
\title{Clean / Separate an address column}
\usage{
clean_address(df, adr_col)
}
\arguments{
\item{df}{dataframe containing addresses}

\item{adr_col}{column containing addresses}
}
\value{
dataframe with cleaned address columns and original address column
}
\description{
Clean / Separate an address column
}
