% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_columns}
\alias{clean_columns}
\title{Clean columns and remove empty columns}
\usage{
clean_columns(df, sort = FALSE, all_character = FALSE)
}
\arguments{
\item{df}{A data frame to clean}

\item{sort}{Sort the columns by name}

\item{all_character}{Change all columns to character type?}
}
\value{
A cleaned dataframe.
}
\description{
Clean columns and remove empty columns
}
