% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data_helpers.R
\name{clean_string}
\alias{clean_string}
\title{Clean a string to Title Case & string squish}
\usage{
clean_string(x, case = FALSE, punct = TRUE, spaces = TRUE)
}
\arguments{
\item{x}{A string to clean}

\item{case}{"lower", "upper", "title", or "title up" for title case with three letter segments in uppercase (useful for many abbreviations).}

\item{punct}{Keep punctuation?}

\item{spaces}{Keep spaces?}
}
\value{
A Title Case String.
}
\description{
Clean a string to Title Case & string squish
}
