% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{cleaner_adr}
\alias{cleaner_adr}
\title{Cleaner Addresses}
\usage{
cleaner_adr(df)
}
\arguments{
\item{df}{dataframe containing addresses and "street_number" & "street_name" columns}
}
\value{
dataframe with cleaned address columns and extra columns removed.
}
\description{
Cleaner Addresses
}
