% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{extract_address_part}
\alias{extract_address_part}
\title{Extract address parts}
\usage{
extract_address_part(x, part = NULL, extract)
}
\arguments{
\item{x}{temporary address column}

\item{part}{Part of the address to extract}

\item{extract}{replace with this vector}
}
\value{
A vector
}
\description{
Extract address parts
}
