% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_register.R
\name{filter_register}
\alias{filter_register}
\title{Filter register data for Shelby County}
\usage{
filter_register(df)
}
\arguments{
\item{df}{Unfiltered register's data}
}
\description{
Filter register data for Shelby County
}
