% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_register.R
\name{get_register}
\alias{get_register}
\title{Get register's data}
\usage{
get_register(vec, field = "address")
}
\arguments{
\item{vec}{Vector of data (for one field) to search}

\item{field}{The field to search. Options:
\itemize{
\item "address" (the default)
\item record_number
\item parcel
}}
}
\value{
dataframe with search results
}
\description{
Get register's data
}
