% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rehab_data.R
\name{get_rehab_data}
\alias{get_rehab_data}
\title{Get rehab data}
\usage{
get_rehab_data(file)
}
\arguments{
\item{file}{The file to call. See shelbyR_raw_paths$file for all}
}
\description{
Get rehab data
}
