% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_rehab_filepath}
\alias{get_rehab_filepath}
\title{Get the filepath for a rehab file/folder}
\usage{
get_rehab_filepath(file = NULL)
}
\arguments{
\item{file}{The file or folder to get the filepath for. Options:
\itemize{
\item \code{NULL} (the default): get the main folder filepath
\item \verb{rehab cases}: get the rehab cases spreadsheet
\item \code{bids}: get the bid days folder
\item \code{neighborly}: get neighborly data in archive folder
\item \code{contractors}: get contractor data
}}
}
\value{
A character string
}
\description{
Get the filepath for a rehab file/folder
}
\examples{
# Get the bid folder
get_rehab_filepath("bids")
}
