% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{install_rehab_folder}
\alias{install_rehab_folder}
\title{Install the filepath of your rehab folder}
\usage{
install_rehab_folder(filepath)
}
\arguments{
\item{filepath}{the filepath to your rehab folder}
}
\description{
Install the filepath of your rehab folder
}
\examples{
\dontrun{
install_rehab_folder("path/to/folder/")
# Reload your environment to use the key without restarting session
readRenviron("~/.Renviron")
# Check the install with:
Sys.getenv("LEAD_REHAB_DIR")
}
}
