% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bids.R
\name{read_bid}
\alias{read_bid}
\title{Get bids from a file.}
\usage{
read_bid(filepath)
}
\arguments{
\item{filepath}{The filepath to get bid information from}
}
\value{
A tibble of bid information and details
}
\description{
Get bids from a file.
}
