% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bids.R
\name{read_bid_details}
\alias{read_bid_details}
\title{Read bid details from a file}
\usage{
read_bid_details(filepath)
}
\arguments{
\item{filepath}{The filepath to get bid details from.}
}
\value{
A tibble of bid details
}
\description{
Read bid details from a file
}
