% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bids.R
\name{read_bid_info}
\alias{read_bid_info}
\title{Read bid information from a file.}
\usage{
read_bid_info(filepath)
}
\arguments{
\item{filepath}{The filepath to get bid information from.}
}
\value{
A tibble of bid information.
}
\description{
Read bid information from a file.
}
