% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bids.R
\name{read_bids}
\alias{read_bids}
\title{Read bids from a certain date.}
\usage{
read_bids(bid_date, nested = TRUE)
}
\arguments{
\item{bid_date}{The date of the bid opening & date of the folder to get bidsheets.}

\item{nested}{Are the bid sheets nested in the folder?}
}
\value{
A table of bid information
}
\description{
Read bids from a certain date.
}
