% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data_helpers.R
\name{remove_address_part}
\alias{remove_address_part}
\title{Remove address parts}
\usage{
remove_address_part(x, part, extract)
}
\arguments{
\item{x}{Temporary address column}

\item{part}{Part of the address to extract}

\item{extract}{replace with this vector}
}
\value{
A vector
}
\description{
Remove address parts
}
