% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelbyR-package.R
\docType{package}
\name{shelbyR-package}
\alias{shelbyR}
\alias{shelbyR-package}
\title{shelbyR: SCDH Helpers}
\description{
Helpers for the SCDH.
}
\author{
\strong{Maintainer}: Sarah Johnson \email{sarahviamemphis@gmail.com}

}
\keyword{internal}
