% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shelbyR_cache_dir}
\alias{shelbyR_cache_dir}
\title{Get cache directory}
\usage{
shelbyR_cache_dir(folder = NULL)
}
\arguments{
\item{folder}{cache folder to return (and create, if needed)}
}
\value{
cache_dir: folder path of cache directory
}
\description{
Get cache directory
}
