% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{shelbyR_get_path}
\alias{shelbyR_get_path}
\title{Get the filepath for a rehab file/folder}
\usage{
shelbyR_get_path(file = NULL)
}
\arguments{
\item{file}{The file or folder to get the filepath for.}
}
\value{
A character string
}
\description{
Get the filepath for a rehab file/folder
}
\examples{
# Get the bid folder
shelbyR_get_path("bids")
}
