% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shelbyR_raw_paths}
\alias{shelbyR_raw_paths}
\title{shelbyR raw data file paths}
\format{
\subsection{\code{shelbyR_raw_paths}}{

A tibble with 11 rows and 5 columns:
\describe{
\item{file}{shorthand for the file to retrieve}
\item{base}{basename of the file}
\item{pattern}{for regex filepaths}
\item{sheet}{excel sheet to pull}
\item{skip}{number of rows to skip for file}
}
}
}
\usage{
shelbyR_raw_paths
}
\description{
shelbyR raw data file paths
}
\keyword{datasets}
