% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shelbyR_read_dataraw}
\alias{shelbyR_read_dataraw}
\title{Read raw Lead + Rehab data into R.}
\usage{
shelbyR_read_dataraw(file)
}
\arguments{
\item{file}{See shelbyR_raw_paths for full list of files}
}
\value{
df of raw data
}
\description{
Read raw Lead + Rehab data into R.
}
