% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shelbyR_set_path}
\alias{shelbyR_set_path}
\title{Install the filepath of your rehab folder}
\usage{
shelbyR_set_path(filepath)
}
\arguments{
\item{filepath}{the filepath to your outlook lead + rehab folder}
}
\description{
Install the filepath of your rehab folder
}
\examples{
\dontrun{
shelbyR_set_path("path/to/folder")
# Reload your environment to use the key without restarting session
readRenviron("~/.Renviron")
# Check the install with:
Sys.getenv("LEAD_REHAB_DIR")
}
}
