% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shelbyR_update_cache}
\alias{shelbyR_update_cache}
\title{Update shelbyR cache data}
\usage{
shelbyR_update_cache(file)
}
\arguments{
\item{file}{Exported contractor neighborly file}
}
\description{
Update shelbyR cache data
}
