% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_register.R
\name{get_register}
\alias{get_register}
\title{Get register's data}
\usage{
get_register(vec, field = "address", cache = TRUE)
}
\arguments{
\item{vec}{Vector of search items for one field (such as a list of addresses).}

\item{field}{Search field to use. Options:
\itemize{
\item \code{address} (the default)
\item \code{record_number}
\item \code{parcel}
}}

\item{cache}{Use cache directory.}
}
\value{
dataframe with search results
}
\description{
Get register's data
}
