% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_rehab_data}
\alias{get_rehab_data}
\title{Get rehab files}
\usage{
get_rehab_data(data = NULL)
}
\arguments{
\item{data}{The files to retrieve
\itemize{
\item \code{neighborly}: Neighborly data in archive
\item \code{contractors}: Contractor doc file
}}
}
\description{
Get rehab files
}
