% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rehab_data.R
\name{get_rehab_data}
\alias{get_rehab_data}
\title{Get rehab data}
\usage{
get_rehab_data(dataset)
}
\arguments{
\item{dataset}{The dataset to call
\itemize{
\item \code{bids}
\item \code{FY23}
\item \code{cancelled}
\item \code{queue}
\item \code{contractors}
\item \verb{neighborly cases}
\item \verb{neighborly logs}
}}
}
\description{
Get rehab data
}
